refactoring

+++ src/main/java/com/ll/gramgram/boundedContext/likeablePerson/service/LikeablePersonService.java
@@ -52,7 +52,7 @@ public class LikeablePersonService {
     }
 
     public LikeablePerson getLikeablePerson(Long id) {
-        Optional<LikeablePerson> likeablePerson = likeablePersonRepository.findById(id.intValue());
+        Optional<LikeablePerson> likeablePerson = likeablePersonRepository.findById(id);
         if (likeablePerson.isPresent()) {
             return likeablePerson.get();
         } else{
